Add ServiceRegister.CheckExistingRelation
ServiceRegister will not check to see if it already contains and instance of a service that meets the given TService.
This is used by the ServiceRegister Methods where the ServiceRegister would otherwise create an instance if a TService was not already registered.

This satisfies a less common but occuring use case where a concrete Service implements more than 1 IService, and the user's expects that Flume is ensuring that Services are 'Singleton' style.
Allows there to be

public interface IMyServiceA { }

public interface IMyServiceB { }

public class MyBigService : IMyServiceA, IMyServiceB
{
    private static int _instanceCount;

    public MyBigService()
    {
        _instanceCount++;

        if (_instanceCount > 1)
            throw new System.Exception($"More than 1 {typeof(MyBigService)} has been created.");
    }
}

container
    .Register<IMyServiceA, MyBigService>()
    .Register<IMyServiceB, MyBigService>();